% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hackathon.vn-package.R
\docType{package}
\name{hackathon.vn-package}
\alias{hackathon.vn}
\alias{hackathon.vn-package}
\title{hackathon.vn: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Vigie-Nature/hackathon.vn}
  \item Report bugs at \url{https://github.com/Vigie-Nature/hackathon.vn/issues}
}

}
\author{
\strong{Maintainer}: Mael Pretet \email{mael.pretet1@mnhn.fr} (\href{https://orcid.org/{{orcid}}}{ORCID}) [copyright holder]

}
\keyword{internal}
